+++ src/textual/_cache.py
@@ -47,6 +47,7 @@ class LRUCache(Generic[CacheKey, CacheValue]):
 
     @property
     def maxsize(self) -> int:
+        """int: Maximum size of cache, before new values evict old values."""
         return self._maxsize
 
     @maxsize.setter
@@ -58,6 +59,14 @@ class LRUCache(Generic[CacheKey, CacheValue]):
 
     def __len__(self) -> int:
         return len(self._cache)
+
+    def grow(self, maxsize: int) -> None:
+        """Grow the maximum size to at least `maxsize` elements.
+
+        Args:
+            maxsize (int): New maximum size.
+        """
+        self.maxsize = max(self.maxsize, maxsize)
 
     def clear(self) -> None:
         """Clear the cache."""